backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
google-generativeai
python-dotenv
sqlalchemy